move struct Metrics to utils
