+++ internal/server/httplayer/update_handlers.go
@@ -25,7 +25,7 @@ func (api *httpAPI) updateJSON(w http.ResponseWriter, r *http.Request) {
 		api.logger.Info("can't read request body", zap.Error(err))
 		return
 	}
-	var metric Metrics
+	var metric utils.Metrics
 	err = json.Unmarshal(body, &metric)
 	if err != nil {
 		http.Error(w, "invalid json data", http.StatusBadRequest)
@@ -147,8 +147,8 @@ func (api *httpAPI) updates(w http.ResponseWriter, r *http.Request) {
 		api.logger.Info("can't read request body", zap.Error(err))
 		return
 	}
-	var metrics []Metrics
-	var appMetrics []applayer.Metrics
+	var metrics []utils.Metrics
+	var appMetrics []utils.Metrics
 	err = json.Unmarshal(body, &metrics)
 
 	if err != nil {
@@ -172,7 +172,7 @@ func (api *httpAPI) updates(w http.ResponseWriter, r *http.Request) {
 			http.Error(w, fmt.Sprintf("%v", applayer.ErrWrongMetricValue), http.StatusBadRequest)
 			return
 		}
-		appMetrics = append(appMetrics, applayer.Metrics(metric))
+		appMetrics = append(appMetrics, metric)
 	}
 
 	errFromAppLayer := api.app.Updates(appMetrics)
