+++ internal/server/storagelayer/storage.go
@@ -24,13 +24,6 @@ type memStorage struct {
 	config  config.Config
 }
 
-type Metrics struct {
-	ID    string   `json:"id"`              // имя метрики
-	MType string   `json:"type"`            // параметр, принимающий значение gauge или counter
-	Delta *int64   `json:"delta,omitempty"` // значение метрики в случае передачи counter
-	Value *float64 `json:"value,omitempty"` // значение метрики в случае передачи gauge
-}
-
 type MemStorage interface {
 	UpdateCounter(string, int64) error
 	UpdateGauge(string, float64) error
@@ -38,7 +31,7 @@ type MemStorage interface {
 	GetGauge(string) (float64, error)
 	GetAll() (map[string]float64, map[string]int64)
 	CheckDBConnection() error
-	Updates(metric []Metrics) error
+	Updates(metric []utils.Metrics) error
 }
 
 var ErrMetricNotFound = errors.New("metric not found")
@@ -138,7 +131,7 @@ func (mem *memStorage) UpdateCounter(metricName string, metricValue int64) (err 
 	return nil
 }
 
-func (mem *memStorage) Updates(metrics []Metrics) (err error) {
+func (mem *memStorage) Updates(metrics []utils.Metrics) (err error) {
 	defer func() {
 		rec := recover()
 		if rec != nil {
@@ -445,7 +438,7 @@ func (mem *memStorage) readFromDB() error {
 	return nil
 }
 
-func (mem *memStorage) updatesWriteToDB(metrics []Metrics) error {
+func (mem *memStorage) updatesWriteToDB(metrics []utils.Metrics) error {
 	db, err := sql.Open("pgx", mem.config.DatabaseDSN)
 	if err != nil {
 		mem.logger.Error("", zap.Error(err))
